requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
cuser==0.0.4
Django==5.0.6
django-cuser==2017.3.16
django-decouple==2.1
django-modeltranslation==0.17.3
django-tinymce==4.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.6
modeltranslation==0.25
multidict==6.0.5
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyserial==3.5
pyTelegramBotAPI==4.16.1
python-decouple==3.8
python-dotenv==0.20.0
python-environ==0.4.54
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
setuptools==70.1.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
telebot==0.0.5
translator==0.0.9
twilio==9.2.2
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
wikipedia==1.4.0
yarl==1.9.4
